test: consolidate 2D array encoding tests into DynamicArrayTest
- Merges tests from Issue1904ReproductionTest and NestedArrayTypeStringTest into DynamicArrayTest as requested by the maintainer.
- Removes the separate test files.
- Refactors the test declarations to use raw types for nested arrays, completely removing the need for unchecked type suppression.
